# openapi/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: AWS Environment Ondemand APIs
  description: |-
    This is an end to end environment ondemand setup api
  contact:
    email: apiteam@swagger.io
  version: 1.0.11
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
paths:
  /application/components:
    get:
      tags:
        - Application
      summary: Get all the components
      description: List all the components.
      operationId: getAllComponents
      parameters:
        - name: components
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Component'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Component'
        '400':
          description: Invalid component
  /aplication:
    get:
      tags:
        - Application
      summary: Get all the applications
      description: List all the applications.
      operationId: getAllApplications
      parameters:
        - name: applications
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '400':
          description: Invalid app
    post:
      tags:
        - Application
      summary: Define a new application releases
      description: Define a new application
      operationId: defineApplication
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              allOf:
                - $ref: '#/components/schemas/Application'
                - type: object
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    put:
      tags:
        - Application
      summary: Update an existing application definition
      description: Update an existing application definition by Id
      operationId: updateAplication
      requestBody:
        description: Update an existent application definition
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Invalid ID supplied
        '404':
          description: Aplication not found
        '422':
          description: Validation exception
  /application/{appId}:
    get:
      tags:
        - Application
      summary: Find application definition by Id
      description: Returns a single application definition
      operationId: getAppDefById
      parameters:
        - name: appId
          in: path
          description: Id of application definition to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Invalid ID supplied
        '404':
          description: Environments not found
    delete:
      tags:
        - Application
      summary: Deletes an application definition
      description: delete a application definition
      operationId: deleteAppDef
      parameters:
        - name: appId
          in: path
          description: Application Definition id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid application definition value
...
components:
  schemas:
    Component:
      type: object
      properties:
        name:
          type: string
          example: 'S3'
        artifactTye:
          type: string
          example: 'Terraform'
        version:
          type: string
          example: '1.1'
    Application:
      type: object
      required:
        - name
        - release
        - components
      properties:
        id:
          type: string
          example: 'uuid'
        name:
          type: string
          example: 'Name'
        release:
          type: string
          example: '1.1'
        components:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Component'
...